added email validation

+++ db/mail.go
@@ -1,6 +1,8 @@
 package db
 
 import (
+	"regexp"
+
 	_ "github.com/mutecomm/go-sqlcipher/v4"
 	"github.com/pkg/errors"
 )
@@ -15,7 +17,11 @@ func (r *Repository) AddEmail(param []string) error {
 		}
 	}
 
-	//stmt, err := r.db.Prepare("delete from email;")
+	if !isEmailValid(param[0]) {
+		return errors.New("Invalid email, not changing settings")
+	}
+
+	// first delete previous settings
 	_, err := r.db.Exec("delete from email;")
 	if err != nil {
 		return errors.Wrap(err, "delete from email")
@@ -39,8 +45,6 @@ func (r *Repository) AddEmail(param []string) error {
 	if err != nil {
 		return errors.Wrap(err, "inserting email")
 	}
-
-	//log.Printf("%s %s added\n", param[0], param[1])
 
 	return nil
 }
@@ -70,104 +74,12 @@ func (r *Repository) GetEmail() (*Email, error) {
 	return &m, nil
 }
 
-/*
-func listDevices() ([]Device, error) {
-	var A []Device
-	rows, err := DBCon.Query("select ArrayType, Cluster, Name, Friendlyname, Username, Password from Arrays")
-	if err != nil {
-		return nil, err
+var emailRegex = regexp.MustCompile("^[a-zA-Z0-9.!#$%&'*+\\/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$")
+
+// isEmailValid checks if the email provided passes the required structure and length.
+func isEmailValid(e string) bool {
+	if len(e) < 3 && len(e) > 254 {
+		return false
 	}
-	defer rows.Close()
-	for rows.Next() {
-		var at DeviceType
-		var Cluster, Name, Friendlyname, Username, Password string
-		err = rows.Scan(&at, &Cluster, &Name, &Friendlyname, &Username, &Password)
-		if err != nil {
-			log.Fatalf("(listedevices) Error in DB %v", err)
-		}
-		var arr1, usr1, psw1 string
-		arr1, err = Decrypt(Name)
-		if err != nil {
-			log.Fatalf("Cannot decrypt array name : %v", err)
-		}
-		usr1, err = Decrypt(Username)
-		if err != nil {
-			log.Fatalf("Cannot decrypt username : %v", err)
-		}
-		psw1, err = Decrypt(Password)
-		if err != nil {
-			log.Fatalf("Cannot decrypt password : %v", err)
-		}
-		A = append(A, Device{at, Cluster, arr1, Friendlyname, usr1, psw1})
-	}
-	err = rows.Err()
-	if err != nil {
-		return nil, err
-	}
-	return A, nil
+	return emailRegex.MatchString(e)
 }
-
-
-func listarrays_names() ([]string, error) {
-	A, err := listDevices()
-	if err != nil {
-		return []string{}, err
-	}
-	var ret []string
-	for _, v := range A {
-		ret = append(ret, v.Type.String()+` `+v.Name+` `+v.Friendlyname+` `+v.Cluster)
-	}
-	return ret, nil
-}
-
-
-
-func getemail() (rcpt_to, mailserver, mailfrom, subject, username, password string, err error) {
-	rows, err := DBCon.Query("select rcpt_to,mailserver,mailfrom,subject,username,password from email")
-	if err != nil {
-		fmt.Println("Select failed")
-		panic(err)
-	}
-	defer rows.Close()
-
-	var rcpt_to1, mailserver1, mailfrom1, subject1, username1, password1 string
-	for rows.Next() {
-		err = rows.Scan(&rcpt_to, &mailserver, &mailfrom, &subject, &username, &password)
-		if err != nil {
-			log.Fatalln(err)
-		}
-
-		rcpt_to1, err = Decrypt(rcpt_to)
-		if err != nil {
-			log.Fatalln(err)
-		}
-		mailserver1, err = Decrypt(mailserver)
-		if err != nil {
-			log.Fatalln(err)
-		}
-		mailfrom1, err = Decrypt(mailfrom)
-		if err != nil {
-			log.Fatalln(err)
-		}
-
-		subject1, err = Decrypt(subject)
-		if err != nil {
-			log.Fatalln(err)
-		}
-		username1, err = Decrypt(username)
-		if err != nil {
-
-			log.Fatalln(err)
-		}
-		password1, err = Decrypt(password)
-		if err != nil {
-			log.Fatalln(err)
-		}
-	}
-	err = rows.Err()
-	if err != nil {
-		log.Fatalln(err)
-	}
-	return rcpt_to1, mailserver1, mailfrom1, subject1, username1, password1
-}
-*/
